+++ includes/SpeechoidConnector.php
@@ -17,13 +17,14 @@ use MediaWiki\MediaWikiServices;
  */
 class SpeechoidConnector {
 
+	/** @var string Speechoid URL, without trailing slash. */
 	private $url;
 
 	public function __construct() {
 		$config = MediaWikiServices::getInstance()
 			->getConfigFactory()
 			->makeConfig( 'wikispeech' );
-		$this->url = $config->get( 'WikispeechSpeechoidUrl' );
+		$this->url = rtrim( $config->get( 'WikispeechSpeechoidUrl' ), '/' );
 	}
 
 	/**
@@ -41,8 +42,7 @@ class SpeechoidConnector {
 		$voice,
 		$text
 	): array {
-		$requestFactory =
-			MediaWikiServices::getInstance()->getHttpRequestFactory();
+		$requestFactory = MediaWikiServices::getInstance()->getHttpRequestFactory();
 		$requestParameters = [
 			'lang' => $language,
 			'voice' => $voice,
@@ -67,13 +67,43 @@ class SpeechoidConnector {
 	}
 
 	/**
-	 * Retrieve default voice setup from Speechoid.
+	 * Retrieve and parse default voices per language from Speechoid.
 	 *
 	 * @since 0.1.5
 	 * @return array Map language => voice
+	 * @throws SpeechoidConnectorException On Speechoid I/O- or JSON parse errors.
 	 */
-	public function listDefaultVoicePerLanguage() {
-		// @todo Awaits implementation in Speechoid Wikispeech-server.
-		return [];
+	public function listDefaultVoicePerLanguage() : array {
+		$defaultVoicesJson = $this->requestDefaultVoices();
+		$status = FormatJson::parse(
+			$defaultVoicesJson,
+			FormatJson::FORCE_ASSOC
+		);
+		if ( !$status->isOK() ) {
+			throw new SpeechoidConnectorException( 'Unexpected response from Speechoid.' );
+		}
+		$defaultVoices = $status->getValue();
+		$defaultVoicePerLanguage = [];
+		foreach ( $defaultVoices as $voice ) {
+			$defaultVoicePerLanguage[ $voice['lang'] ] = $voice['default_voice'];
+		}
+		return $defaultVoicePerLanguage;
 	}
+
+	/**
+	 * Retrieve default voices par language from Speechoid
+	 *
+	 * @since 0.1.6
+	 * @return string JSON response
+	 * @throws SpeechoidConnectorException On Speechoid I/O error.
+	 */
+	public function requestDefaultVoices(): string {
+		$requestFactory = MediaWikiServices::getInstance()->getHttpRequestFactory();
+		$responseString = $requestFactory->get( $this->url . '/default_voices' );
+		if ( !$responseString ) {
+			throw new SpeechoidConnectorException( 'Unable to communicate with Speechoid.' );
+		}
+		return $responseString;
+	}
+
 }